A cloud based image editing Android Application
This is an android application that ask for an imge from local device
and attach graphical component onto it. User can edit the image and
upload it to the cloud side and pull it back after log in.

This is as showcase of combination of transform, color rendering and
ajax in Android platform. The compatibility of this application is from
Android 2.3 to 4.4.2. The cloud side is build with PHP and mySQL.